refactor(ui): extract voice bring-up engine from the fat route handler
Move the Docker/compose bring-up infrastructure out of the ~1041-line
routes/admin/voice/+server.ts into lib/server/voice/bring-up.ts, mirroring
the lib/server/endpoints.ts <-> admin/endpoints/+server.ts thin-route split.

The service now owns: image inspection (dockerImagePresent/resolveServiceImage),
the TCP port probe, container-health + rootless/nvidia-runtime detection,
compose-overlay generation (CDI + rootless fallbacks), the in-memory
background-job registry (activeJobs), the compose-up + 180s /health poll
lifecycle, and the engageVoiceAddon orchestration. The route keeps only
request validation, the openpalm-voice preset fill, and mapping the typed
VoiceEngageResult to HTTP responses. Behavior is identical (same endpoints,
responses, status codes, and job semantics).

Overlay YAML builders are split into pure buildCdiOverlayYaml/
buildRootlessOverlayYaml functions and covered, along with isLargeImageTag,
resolveDefaultProfile, and the job-registry state transitions, by a new
bring-up.vitest.ts (12 tests). The route dropped from ~1041 to ~300 lines.

svelte-check: 0 errors. Existing server.vitest.ts (18 tests) still green.
